+++ src/app/ring-with-text/ring-with-text.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-ring-with-text',
+  templateUrl: './ring-with-text.component.html',
+  styleUrls: ['./ring-with-text.component.css']
+})
+export class RingWithTextComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}