+++ src/lib/index.ts
@@ -1,5 +1,4 @@
 export { ActionPanelComponent } from './controls/action-panel/action-panel.component';
-export { ActionPanelJqeuryHelper } from './controls/action-panel/actionPanelJqueryHelper';
 export { TextareaComponent } from './controls/textarea/textarea.component';
 export { ButtonComponent } from './controls/button/button.component';
 export { CardComponent } from './controls/card/card.component';
@@ -31,7 +30,6 @@ export { ListColumnComponent } from './controls/list/list-column.component';
 export { ListColumnCheckboxComponent } from './controls/list/list-column-checkbox.component';
 export { ListColumnTrashcanComponent } from './controls/list/list-column-trashcan.component';
 export { ListItemComponent } from './controls/list-item/list-item.component';
-export { ListItemJqeuryHelper } from './controls/list-item/listItemJqueryHelper';
 export { MenuComponent } from './controls/menu/menu.component';
 export { MenuItemComponent } from './controls/menu/menu-item.component';
 export { MenuSeparatorComponent } from './controls/menu/menu-separator.component';
@@ -52,6 +50,7 @@ export { ModalContentComponent } from './controls/modal/modal-content.component'
 export { ModalHeaderComponent } from './controls/modal/modal-header.component';
 export { ModalFooterComponent } from './controls/modal/modal-footer.component';
 export { CloseButtonComponent } from './controls/close-button/close-button.component';
+export { RingWithTextComponent } from './controls/ring-with-text/ring-with-text.component';
 export { TableComponent } from './controls/table/table.component';
 export { TableRowColumnComponent } from './controls/table/table-row-column.component';
 export { TableHeaderColumnComponent } from './controls/table/table-header-column.component';
@@ -72,6 +71,7 @@ export { IMenu, IMenuGroup, IMenuItem } from './models/menu.model';
 export { NotificationIcon } from './models/notificationIcon.model';
 export { NotificationType } from './models/notificationType.model';
 export { RowNotification } from './models/rowNotification.model';
+export { Notification } from './models/notification.model';
 export { ISelectableItem, SelectableItem } from './models/selectableItem.model';
 
 export { DropdownItemToSelectedTextPipe } from './pipes/dropdownItemToSelectedTextPipe';